Using __init__.py instead of api.py
